+++ odoo/local-src/specific_payment_mode/models/account_payment_method.py
@@ -2,7 +2,9 @@
 # Copyright 2017 Camptocamp SA
 # License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl).
 
-from odoo import models, fields, api
+from odoo.exceptions import UserError
+
+from odoo import models, fields, api, _
 
 
 class AccountPaymentMethod(models.Model):
@@ -32,13 +34,22 @@ class AccountPaymentMethod(models.Model):
     @api.multi
     def generate_payment_modes(self):
         self.ensure_one()
+
+        if self.env.user.company_id != self.env.ref('base.main_company'):
+            raise UserError(
+                _('You must be connected '
+                  'on "Depil Tech Holding" company to generate payment modes.')
+            )
+
         wizard_model = self.env['account.payment.mode.generator']
 
         values = {
             'account_payment_method_id': self.id,
         }
         values.update(wizard_model.get_default_values(self))
-        wizard = wizard_model.create(values)
+        # We must be a sudo here, because except admin user,
+        # some objects (journals) are readable only for the connected company
+        wizard = wizard_model.sudo().create(values)
 
         action_data = self.env.ref(
             'specific_payment_mode.account_payment_mode_generator_action'